android progress

+++ talosblockchainj/BlockAdit/app/src/main/java/ch/ethz/blockadit/blockadit/model/DataEntryAgrDate.java
@@ -41,5 +41,5 @@ public interface DataEntryAgrDate {
 
     public Date getDate();
 
-    public int getValue();
+    public double getValue();
 }
